moved safe_iter to util module and more descriptive name
